Extract JavaScript from PHP
fixes #84
