+++ src/fuer_einsteiger_l.php
@@ -10,49 +10,6 @@ require_once __DIR__.'/model/User.php';
 require_once __DIR__.'/model/UserRepository.php';
 require_once __DIR__.'/components/users/olz_user_info_card/olz_user_info_card.php';
 
-?>
-
-<script type='text/javascript'>
-
-function highlight_menu(page) {
-    var menuContainerElem = document.getElementById('menu-container');
-    var menuContainerStyle = window.getComputedStyle(menuContainerElem);
-    var menuContainerOpacity = menuContainerStyle.getPropertyValue('opacity');
-    if (menuContainerOpacity < 0.5) {
-        return;
-    }
-    var elem = document.getElementById("menu_a_page"+page);
-    var rect = elem.getBoundingClientRect();
-    var pointer = document.createElement("img");
-    pointer.style.pointerEvents = "none";
-    pointer.style.position = "fixed";
-    pointer.style.zIndex = "1000";
-    pointer.style.top = (rect.top+rect.height/2-50)+"px";
-    pointer.style.left = (rect.left+rect.width)+"px";
-    pointer.style.height = "100px";
-    pointer.style.border = "0px";
-    pointer.src = "icns/arrow_red.svg";
-    pointer.id = "highlight_menu_"+page;
-    document.documentElement.appendChild(pointer);
-    window.setTimeout("highlight_menu_ani("+page+", 0)", 100);
-}
-function highlight_menu_ani(page, step) {
-    if (step==8) step = 0;
-    var elem = document.getElementById("highlight_menu_"+page);
-    if (!elem) return;
-    elem.style.marginLeft = (Math.sin(step*2*3.1415/8)*4)+"px";
-    window.setTimeout("highlight_menu_ani("+page+", "+(step+1)+")", 100);
-}
-function unhighlight_menu(page) {
-    var elem = document.getElementById("highlight_menu_"+page);
-    if (elem) {
-        elem.parentElement.removeChild(elem);
-    }
-}
-</script>
-
-<?php
-
 $role_repo = $entityManager->getRepository(Role::class);
 $buessli_role = $role_repo->findOneBy(['username' => 'buessli']);
 $aktuariat_role = $role_repo->findOneBy(['username' => 'aktuariat']);
@@ -159,5 +116,3 @@ foreach ($buessli_assignees as $buessli_assignee) {
     echo olz_user_info_card($buessli_assignee);
 }
 echo "</div>";
-
-?>
