+++ src/header.php
@@ -220,30 +220,18 @@ $are_kids_winners = ($percent_kids > $percent_j_und_s);
 $color_kids = $are_kids_winners ? 'rgb(0,100,0)' : 'rgb(180,0,0)';
 $color_j_und_s = $are_kids_winners ? 'rgb(180,0,0)' : 'rgb(0,100,0)';
 
-echo "<script>";
-echo "function toggle(ident) {";
-echo "var elem = document.getElementById(ident);";
-echo "var isShown = (elem.style.display == 'block');";
-echo "document.getElementById('ranking-kids-2020').style.display = 'none';";
-echo "document.getElementById('ranking-kids-2019').style.display = 'none';";
-echo "document.getElementById('ranking-junds-2020').style.display = 'none';";
-echo "document.getElementById('ranking-junds-2019').style.display = 'none';";
-echo "elem.style.display = (isShown ? 'none' : 'block');";
-echo "return false;";
-echo "}";
-echo "</script>";
 echo "<div class='header-box'><div style='width:80px;' class='box_ganz'><div style='display: flow-root; width:80px; border:0px;'>";
 echo "<h2 style='font-size:12px; border: 0; padding-left: 0; text-align:center; margin-top: 0;'>JOM-Jugend</h2>";
 echo "<div style='margin-top:-4px; font-size:18px; text-align:center; color:".$color_kids.";' title='Wie viele Jugndliche im Vergleich zu 2019'>".round($percent_kids, 2)."%</div>";
 echo "<div style='margin-top:-2px; font-size:12px; text-align:center;' title='Anzahl Starts von Jugndlichen 2020 / 2019'>";
-echo "<a href='javascript:' onclick='toggle(&quot;ranking-kids-2020&quot;); return false;' style='color:".$color_kids.";'>{$starts_kids_2020}</a> / ";
-echo "<a href='javascript:' onclick='toggle(&quot;ranking-kids-2019&quot;); return false;' style='color:".$color_kids.";'>{$starts_kids_2019}</a>";
+echo "<a href='javascript:' onclick='headerToggle(&quot;ranking-kids-2020&quot;); return false;' style='color:".$color_kids.";'>{$starts_kids_2020}</a> / ";
+echo "<a href='javascript:' onclick='headerToggle(&quot;ranking-kids-2019&quot;); return false;' style='color:".$color_kids.";'>{$starts_kids_2019}</a>";
 echo "</div>";
 echo "<h2 style='margin-top:2px; font-size:12px; border:0px; padding-left: 0; text-align:center;'>J&amp;S-Leiter</h2>";
 echo "<div style='margin-top:-4px; font-size:18px; text-align:center; color:".$color_j_und_s.";' title='Wie viele J&S-Leiter im Vergleich zu 2019'>".round($percent_j_und_s, 2)."%</div>";
 echo "<div style='margin-top:-2px; font-size:12px; text-align:center;' title='Anzahl Starts von OLZ-J&S-Leitern 2020 / 2019'>";
-echo "<a href='javascript:' onclick='toggle(&quot;ranking-junds-2020&quot;); return false;' style='color:".$color_j_und_s.";'>{$starts_j_und_s_2020}</a> / ";
-echo "<a href='javascript:' onclick='toggle(&quot;ranking-junds-2019&quot;); return false;' style='color:".$color_j_und_s.";'>{$starts_j_und_s_2019}</a>";
+echo "<a href='javascript:' onclick='headerToggle(&quot;ranking-junds-2020&quot;); return false;' style='color:".$color_j_und_s.";'>{$starts_j_und_s_2020}</a> / ";
+echo "<a href='javascript:' onclick='headerToggle(&quot;ranking-junds-2019&quot;); return false;' style='color:".$color_j_und_s.";'>{$starts_j_und_s_2019}</a>";
 echo "</div>";
 echo "</div></div></div>";
 echo $htmlout_kids_2019;
