+++ src/kontakt_d.php
@@ -6,44 +6,6 @@
 
 require_once __DIR__.'/config/paths.php';
 require_once __DIR__.'/config/database.php';
-
-echo "<script type='text/javascript'>
-var highlighttimer = false;
-function highlight_organigramm(id) {
-    highlight_organigramm_scroll(id);
-}
-function highlight_organigramm_scroll(id) {
-    var elem = document.getElementById(id);
-    if (/box\\-[0-9]+\\-[0-9]+/.exec(elem.parentElement.id)) elem = elem.parentElement;
-    elem.style.backgroundColor = \"rgba(0,0,0,0)\";
-    var rect = elem.getBoundingClientRect();
-    var optimalPageYOffset = window.pageYOffset+rect.top+rect.height/2-window.innerHeight/2;
-    var nextPageYOffset = window.pageYOffset+(optimalPageYOffset-window.pageYOffset)/4;
-    if (nextPageYOffset<=0) {
-        window.scrollTo(0, 0);
-        highlight_organigramm_color(id)
-    } else if (document.getElementsByTagName(\"body\")[0].offsetHeight-window.innerHeight<=nextPageYOffset) {
-        window.scrollTo(0, document.getElementsByTagName(\"body\")[0].offsetHeight-window.innerHeight);
-        highlight_organigramm_color(id)
-    } else if (Math.abs(nextPageYOffset-optimalPageYOffset)<=3) {
-        window.scrollTo(0, optimalPageYOffset);
-        highlight_organigramm_color(id)
-    } else {
-        window.scrollTo(0, Math.round(nextPageYOffset));
-        window.setTimeout(function () {highlight_organigramm_scroll(id);}, 50);
-    }
-}
-function highlight_organigramm_color(id) {
-    var elem = document.getElementById(id);
-    if (/box\\-[0-9]+\\-[0-9]+/.exec(elem.parentElement.id)) elem = elem.parentElement;
-    for (var i=0; i<20; i++) {
-        window.setTimeout((function (i) {return function () {
-            elem.style.backgroundColor = \"rgba(0,220,0,\"+Math.pow(Math.sin(i*Math.PI/12), 2)+\")\";
-        };})(i), i*100);
-    }
-}
-</script>";
-
 require_once __DIR__.'/model/Role.php';
 require_once __DIR__.'/model/RoleRepository.php';
 require_once __DIR__.'/model/User.php';
@@ -98,40 +60,3 @@ echo "<div id='organigramm'><h2>Häufig gesucht</h2>
 <div><b><a href='javascript:highlight_organigramm(&quot;link-role-19&quot;)' class='linkint'>Kleiderverkauf</a></b></div>
 <div><b>PC-Konto: 85-256448-8</b></div>
 <h2>Organigramm OL Zimmerberg</h2>".$org."</div>";
-
-echo "<script type='text/javascript'>
-function olz_marquee(elem) {
-    var om = elem.getAttribute(\"olzmarquee\");
-    if (om) {
-        var subdiv = document.createElement(\"div\");
-        var subspan = document.createElement(\"span\");
-        subspan.innerHTML = elem.innerHTML;
-        elem.innerHTML = \"\";
-        elem.appendChild(subdiv);
-        subdiv.style.textAlign = \"inherit\";
-        subdiv.style.width = subdiv.offsetWidth+\"px\";
-        subdiv.style.overflowX = \"hidden\";
-        subdiv.style.whiteSpace = \"nowrap\";
-        subdiv.appendChild(subspan);
-        if (subdiv.offsetWidth<subspan.offsetWidth) {
-            var sw = subspan.offsetWidth-subdiv.offsetWidth;
-            window.setTimeout((function (subdiv, sw) {return function () {
-                if (subdiv.scrollLeft<sw) {
-                    subdiv.scrollLeft += 1;
-                    window.setTimeout(arguments.callee, 75);
-                } else {
-                    window.setTimeout((function (subdiv, cl) {return function () {
-                        subdiv.scrollLeft = 0;
-                        window.setTimeout(cl, 1500);
-                    };})(subdiv, arguments.callee), 1500);
-                }
-            };})(subdiv, sw), 100);
-        }
-    }
-    var cld = elem.children;
-    for (var i=0; i<cld.length; i++) {
-        olz_marquee(cld[i]);
-    }
-}
-//olz_marquee(document.getElementById(\"organigramm\"));
-</script>";
