+++ src/trophy_d.php
@@ -12,19 +12,6 @@
 <p style='text-align:center;'>Die Versicherung ist Sache der Teilnehmenden. Der Veranstalter lehnt, soweit gesetzlich zulässig, jede Haftung ab.</p>
 
 <h3 style='font-size:18px;'>Etappen</h3>
-<script type="text/javascript" src="library/wgs84_ch1903/wgs84_ch1903.js"></script>
-<script type="text/javascript">
-function map(xkoord, ykoord) {
-    var breite = 400;
-
-    // Neue Mapbox Karte
-    var lat = CHtoWGSlat(xkoord, ykoord);
-    var lng = CHtoWGSlng(xkoord, ykoord);
-
-    // Link (im Moment wird noch auf Search.ch verlinkt, denn dort sieht man öV Haltestellen)
-    return "<a href='http://map.search.ch/"+xkoord+","+ykoord+"' target='_blank'><img src='https://api.mapbox.com/styles/v1/allestuetsmerweh/ckgf9qdzm1pn319ohqghudvbz/static/pin-l+009000("+lng+","+lat+")/"+lng+","+lat+",13,0/"+breite+"x300?access_token=pk.eyJ1IjoiYWxsZXN0dWV0c21lcndlaCIsImEiOiJHbG9tTzYwIn0.kaEGNBd9zMvc0XkzP70r8Q' class='noborder test-flaky' style='margin:0px;padding:0px;align:center;border:1px solid #000000;'><\/a>";
-}
-</script>
 
 <?php
 
@@ -94,7 +81,7 @@ for ($i = 0; $i < count($etappen); $i++) {
     <tr><td></td><td><a href='?page=3#id".$etappe[13]."' class='linkint'>Termine-Eintrag</a>".($etappe[9] ? "</td></tr>
     <tr><td></td><td><a href='".$etappe[9]."' class='linkext'>weitere Infos</a>" : "").($etappe[12] && is_file("{$data_path}results/{$etappe[12]}.xml") ? "</td></tr>
     <tr><td></td><td><a href='{$code_href}resultate/?file=".$etappe[12].".xml' class='linkint'>Resultate</a>" : "")."</td></tr>
-    </table></div></td><td style='width:20%; padding:5px 0px 5px 10px;'>".($etappe[4] != 0 ? "<script>document.write(map(".$etappe[4].",".$etappe[5]."))</script>" : "")."</td></tr>";
+    </table></div></td><td style='width:20%; padding:5px 0px 5px 10px;'>".($etappe[4] != 0 ? "<script>document.write(getMapHtml(".$etappe[4].",".$etappe[5]."))</script>" : "")."</td></tr>";
     if (isset($_SESSION['auth']) && $_SESSION['auth'] == 'all' && $etappe[12]) {
         if (isset($_FILES["resultate_upload_".$etappe[13]])) {
             move_uploaded_file(
